chore(maven-publish.yml): update maven publish workflow to use correct path for settings.xml file
The maven-publish.yml workflow file has been updated to fix the path for the settings.xml file used during the deployment step. The previous path was incorrect and has been replaced with the correct path using the $HOME/$GITHUB_WORKSPACE variable. This ensures that the correct settings.xml file is used for the deployment process.