fix(maven-publish.yml): update build and publish steps to use correct working directory and environment variables
The build step in the Maven workflow was updated to use the correct working directory and run the `mvn` command with the correct `pom.xml` file located in the `trial01` directory.

The publish step was also updated to use the correct working directory and environment variables. The `mvn` command now includes the `-s` flag to specify the path to the `settings.xml` file located in the root of the repository. The `GITHUB_TOKEN` environment variable is also set to the value of the `github.token` secret.

Additionally, the `deploy-trial02` job was removed as it is no longer needed.