chore(maven-publish.yml): add workflow for deploying trial02 to GitHub Packages Apache Maven
This commit adds a new workflow to the `.github/workflows/maven-publish.yml` file. The workflow is named `deploy-trial02` and is triggered on the `ubuntu-latest` runner. It has permissions to read contents and write packages.

The workflow consists of the following steps:
1. Checkout the repository using `actions/checkout@v3`.
2. Set up JDK 11 using `actions/setup-java@v3`.
3. Build the project with Maven using the `mvn -B package --file trial02/pom.xml` command.
4. Publish the built artifact to GitHub Packages Apache Maven using the `mvn -DaltDeploymentRepository=github::default::https://maven.pkg.github.com/DaichiMiyajima/platform-openapi-action-spring deploy -f trial02/pom.xml -s $GITHUB_WORKSPACE/settings.xml` command.

The workflow is configured to use the `GITHUB_TOKEN` secret for authentication.